requirements.txt
 Django>=5.0,<6.0
 google-generativeai>=0.8.0
 python-dotenv>=1.0.1


